+++ app/src/main/AndroidManifest.xml
@@ -42,7 +42,7 @@
         android:logo="@drawable/videos_by_google_banner"
         android:theme="@style/Theme.Example.Leanback" >
         <activity
-            android:name="com.example.android.tvleanback.MainActivity"
+            android:name=".ui.MainActivity"
             android:icon="@drawable/videos_by_google_banner"
             android:label="@string/app_name"
             android:logo="@drawable/videos_by_google_banner"
@@ -56,7 +56,7 @@
         </activity>
 
         <activity
-            android:name="com.example.android.tvleanback.DetailsActivity"
+            android:name=".ui.MovieDetailsActivity"
             android:exported="true">
 
             <!-- Receives the search request. -->
@@ -70,27 +70,27 @@
                 android:resource="@xml/searchable" />
         </activity>
 
-        <activity android:name="com.example.android.tvleanback.PlaybackOverlayActivity"
+        <activity android:name=".ui.PlaybackOverlayActivity"
             android:exported="true" />
 
         <activity
-            android:name="com.example.android.tvleanback.VerticalGridActivity"
+            android:name=".ui.VerticalGridActivity"
             android:exported="true"
-            android:parentActivityName="com.example.android.tvleanback.MainActivity" />
+            android:parentActivityName=".ui.MainActivity" />
         <activity
-            android:name="com.example.android.tvleanback.SearchActivity"
+            android:name=".ui.SearchActivity"
             android:exported="true" />
 
-        <activity android:name="com.example.android.tvleanback.BrowseErrorActivity"
+        <activity android:name=".ui.BrowseErrorActivity"
             android:exported="true" />
 
         <!-- Provides search suggestions for keywords against video meta data. -->
-        <provider android:name="com.example.android.tvleanback.VideoContentProvider"
+        <provider android:name=".data.VideoContentProvider"
             android:authorities="com.example.android.tvleanback"
             android:exported="true" />
 
         <receiver
-            android:name="com.example.android.tvleanback.BootupActivity"
+            android:name=".recommendation.BootupActivity"
             android:enabled="true"
             android:exported="false" >
             <intent-filter>
@@ -99,7 +99,7 @@
         </receiver>
 
         <service
-            android:name="com.example.android.tvleanback.UpdateRecommendationsService"
+            android:name=".recommendation.UpdateRecommendationsService"
             android:enabled="true" />
     </application>
