+++ app/src/main/java/com/example/android/tvleanback/data/VideoDatabase.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.data;
 
 import android.app.SearchManager;
 import android.content.ContentValues;
@@ -26,6 +26,9 @@ import android.database.sqlite.SQLiteQueryBuilder;
 import android.media.Rating;
 import android.provider.BaseColumns;
 import android.util.Log;
+
+import com.example.android.tvleanback.R;
+import com.example.android.tvleanback.model.Movie;
 
 import org.json.JSONException;
