+++ app/src/main/java/com/example/android/tvleanback/data/VideoItemLoader.java
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.data;
 
 import android.content.AsyncTaskLoader;
 import android.content.Context;
 import android.util.Log;
+
+import com.example.android.tvleanback.model.Movie;
 
 import java.util.HashMap;
 import java.util.List;
