+++ app/src/main/java/com/example/android/tvleanback/model/Movie.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.model;
 
 import android.util.Log;
