+++ app/src/main/java/com/example/android/tvleanback/presenter/CardPresenter.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.presenter;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -24,6 +24,9 @@ import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.android.tvleanback.R;
+import com.example.android.tvleanback.Utils;
+import com.example.android.tvleanback.model.Movie;
 import com.squareup.picasso.Picasso;
 import com.squareup.picasso.Target;
