+++ app/src/main/java/com/example/android/tvleanback/recommendation/RecommendationBuilder.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.recommendation;
 
 import android.app.Notification;
 import android.app.NotificationManager;
@@ -23,6 +23,8 @@ import android.os.Bundle;
 import android.support.v4.app.NotificationCompat;
 import android.util.Log;
 
+import com.example.android.tvleanback.R;
+import com.example.android.tvleanback.Utils;
 import com.squareup.picasso.Picasso;
 
 import java.io.IOException;
