+++ app/src/main/java/com/example/android/tvleanback/ui/ErrorFragment.java
@@ -11,11 +11,13 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.ui;
 
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
+
+import com.example.android.tvleanback.R;
 
 /*
  * This class demonstrates how to extend ErrorFragment
