+++ app/src/main/java/com/example/android/tvleanback/ui/MainActivity.java
@@ -12,24 +12,26 @@
  * the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.ui;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 
+import com.example.android.tvleanback.R;
+
 /*
- * VerticalGridActivity that loads VerticalGridFragment
+ * MainActivity class that loads MainFragment
  */
-public class VerticalGridActivity extends Activity {
+public class MainActivity extends Activity {
     /**
      * Called when the activity is first created.
      */
+
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.vertical_grid);
-        getWindow().setBackgroundDrawableResource(R.drawable.grid_bg);
+        setContentView(R.layout.main);
     }
 
     @Override
