+++ app/src/main/java/com/example/android/tvleanback/ui/MovieDetailsFragment.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.ui;
 
 import android.content.Intent;
 import android.graphics.Bitmap;
@@ -21,7 +21,6 @@ import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v17.leanback.app.BackgroundManager;
-import android.support.v17.leanback.app.DetailsFragment;
 import android.support.v17.leanback.widget.Action;
 import android.support.v17.leanback.widget.ArrayObjectAdapter;
 import android.support.v17.leanback.widget.ClassPresenterSelector;
@@ -41,6 +40,13 @@ import android.util.DisplayMetrics;
 import android.util.Log;
 import android.widget.Toast;
 
+import com.example.android.tvleanback.PicassoBackgroundManagerTarget;
+import com.example.android.tvleanback.R;
+import com.example.android.tvleanback.Utils;
+import com.example.android.tvleanback.data.VideoProvider;
+import com.example.android.tvleanback.model.Movie;
+import com.example.android.tvleanback.presenter.CardPresenter;
+import com.example.android.tvleanback.presenter.DetailsDescriptionPresenter;
 import com.squareup.picasso.Picasso;
 import com.squareup.picasso.Target;
 
@@ -54,7 +60,7 @@ import java.util.Map;
  * LeanbackDetailsFragment extends DetailsFragment, a Wrapper fragment for leanback details screens.
  * It shows a detailed view of video and its meta plus related videos.
  */
-public class LeanbackDetailsFragment extends DetailsFragment {
+public class MovieDetailsFragment extends android.support.v17.leanback.app.DetailsFragment {
     private static final String TAG = "DetailsFragment";
 
     private static final int ACTION_WATCH_TRAILER = 1;
@@ -89,12 +95,12 @@ public class LeanbackDetailsFragment extends DetailsFragment {
         mMetrics = new DisplayMetrics();
         getActivity().getWindowManager().getDefaultDisplay().getMetrics(mMetrics);
 
-        mSelectedMovie = (Movie) getActivity().getIntent().getSerializableExtra(DetailsActivity.MOVIE);
+        mSelectedMovie = (Movie) getActivity().getIntent().getSerializableExtra(MovieDetailsActivity.MOVIE);
         if (null != mSelectedMovie || checkGlobalSearchIntent()) {
             Log.d(TAG, "DetailsActivity movie: " + mSelectedMovie.toString());
             mDetailRowBuilderTask = (DetailRowBuilderTask) new DetailRowBuilderTask().execute(mSelectedMovie);
             mDorPresenter.setSharedElementEnterTransition(getActivity(),
-                    DetailsActivity.SHARED_ELEMENT_NAME);
+                    MovieDetailsActivity.SHARED_ELEMENT_NAME);
             updateBackground(mSelectedMovie.getBackgroundImageURI());
             setOnItemViewClickedListener(new ItemViewClickedListener());
         } else {
@@ -197,7 +203,7 @@ public class LeanbackDetailsFragment extends DetailsFragment {
                 public void onActionClicked(Action action) {
                     if (action.getId() == ACTION_WATCH_TRAILER) {
                         Intent intent = new Intent(getActivity(), PlaybackOverlayActivity.class);
-                        intent.putExtra(DetailsActivity.MOVIE, mSelectedMovie);
+                        intent.putExtra(MovieDetailsActivity.MOVIE, mSelectedMovie);
                         startActivity(intent);
                     } else {
                         Toast.makeText(getActivity(), action.toString(), Toast.LENGTH_SHORT).show();
@@ -242,13 +248,13 @@ public class LeanbackDetailsFragment extends DetailsFragment {
             if (item instanceof Movie) {
                 Movie movie = (Movie) item;
                 Log.d(TAG, "Item: " + item.toString());
-                Intent intent = new Intent(getActivity(), DetailsActivity.class);
-                intent.putExtra(DetailsActivity.MOVIE, movie);
+                Intent intent = new Intent(getActivity(), MovieDetailsActivity.class);
+                intent.putExtra(MovieDetailsActivity.MOVIE, movie);
 
                 Bundle bundle = ActivityOptionsCompat.makeSceneTransitionAnimation(
                         getActivity(),
                         ((ImageCardView) itemViewHolder.view).getMainImageView(),
-                        DetailsActivity.SHARED_ELEMENT_NAME).toBundle();
+                        MovieDetailsActivity.SHARED_ELEMENT_NAME).toBundle();
                 getActivity().startActivity(intent, bundle);
             }
         }
