+++ app/src/main/java/com/example/android/tvleanback/ui/SearchActivity.java
@@ -12,24 +12,25 @@
  * the License.
  */
 
-package com.example.android.tvleanback;
+package com.example.android.tvleanback.ui;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 
+import com.example.android.tvleanback.R;
+
 /*
- * MainActivity class that loads MainFragment
+ * SearchActivity for SearchFragment
  */
-public class MainActivity extends Activity {
+public class SearchActivity extends Activity {
     /**
      * Called when the activity is first created.
      */
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.main);
+        setContentView(R.layout.search);
     }
 
     @Override
