+++ app/src/main/res/layout/playback_controls.xml
@@ -32,7 +32,7 @@
 
     <fragment
         android:id="@+id/playback_controls_fragment"
-        android:name="com.example.android.tvleanback.PlaybackOverlayFragment"
+        android:name="com.example.android.tvleanback.ui.PlaybackOverlayFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
