+++ app/src/main/res/layout/search.xml
@@ -16,7 +16,7 @@
 -->
 
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
-          android:name="com.example.android.tvleanback.SearchFragment"
+          android:name="com.example.android.tvleanback.ui.SearchFragment"
           android:id="@+id/search_fragment"
           android:layout_width="match_parent"
           android:layout_height="match_parent"
